requirements.txt
streamlit
yfinance
pandas_ta
plotly
requests
gTTS